third_party/blink/web_tests/flag-specific/highdpi/fast/forms/select-popup/popup-menu-scrollbar-button-scrolls-expected.txt

This is a testharness.js-based test.
FAIL Scrollbar clicks in a popup must scroll assert_greater_than: expected a number greater than 0 but got 0
Harness: the test ran to completion.
